JSON file for input

+++ src/main/java/com/ailo/zombie/apocalypse/dto/Coordinates.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Coordinates {
-    long x;
-    long y;
+    int x;
+    int y;
 }
